Simplify 2 vertex Evolution Operator
For the case of just 2 vertices it is possible to massively simplify
the evolution operator by solving the reduced problem in the quantum number
sector corresponding to the initial state. This leads to a much faster and more
efficient computation of the bond evolution operator.

At this stage the code still works by relying on TEBD routines. However further optimisation is possible with 2 vertex specific code.