Branch for finding and fixing dynamics bug
The dynamics code has revealed a bug in the TEBD routine.
This leads to asymmetric answers, even for just two vertices, and
an incorrect sign in some places.
When there is no coupling between sites the answers are correct,
but when there is a coupling but no single vertex terms the answer
is wrong.